requirements.txt
fastembed-gpu
qdrant-client
python-dotenv
pandas

transformers 

torch
